Separate out error handling display logic with new function createArgvInterface
